move composer files to DTO and Enum namespaces
