+++ lib/Composer/ComposerConfigProvider.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Einenlum\PhpStackDetector\Composer;
 
+use Einenlum\PhpStackDetector\DTO\Composer\ComposerConfig;
+use Einenlum\PhpStackDetector\DTO\Enum\ComposerConfigType;
 use Einenlum\PhpStackDetector\DirectoryCrawler\AdapterInterface;
 use Einenlum\PhpStackDetector\Exception\CacheMissException;
 use Einenlum\PhpStackDetector\Exception\ResourceNotFoundException;
