+++ lib/DTO/Enum/ComposerConfigType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Einenlum\PhpStackDetector\Composer;
+namespace Einenlum\PhpStackDetector\DTO\Enum;
 
 enum ComposerConfigType: string
 {
