+++ lib/PhpConfigurationDetector.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Einenlum\PhpStackDetector;
 
-use Einenlum\PhpStackDetector\Composer\ComposerConfig;
+use Einenlum\PhpStackDetector\DTO\Composer\ComposerConfig;
 use Einenlum\PhpStackDetector\Composer\ComposerConfigProvider;
-use Einenlum\PhpStackDetector\Composer\ComposerConfigType;
+use Einenlum\PhpStackDetector\DTO\Enum\ComposerConfigType;
 use Einenlum\PhpStackDetector\DTO\PhpConfiguration;
 use Einenlum\PhpStackDetector\DTO\PhpVersion;
